Pf 215 do not log method return values (#10)
* Re-export @NoLog

* Bump version

* Don't log method return values

Prevent `@LogMethodCall` from logging sensitive information. This is a breaking change as it changes the existing behavior.

* Add parameter to log return value

Adds a paramter to `@LogMethodCall` to log the return value if desired.

* Remove unused inject decorators

* Fix `@LogMethoCall` not logging falsy values

* Split logResult into two methods

so that it uses less optional parameters

* Remove loggerField parameter

loggerField can only ever be 'logger' (as specified by the type), so it seems unnecessary to have it as a parameter

* Log return value by default

When using this I noticed that `@LogMethodCall` is awkward to use with a default behavior of not logging the return value. This is especially true because argument logging works by logging by default, and only hiding the argument if `@NoLog` is specified.

* Remove redundant words in test description

* Formatting

* Fix eslint error

Fixes missing return type warning
Fixes "Don't use `Object` as a type"